moving travel domain under domains folder
